managed to create a setup.py for get_seq which allows for bash auto completion and utilisation; so instead of: 	 python ~/path/to/package/get_seq.py <input> we do: 	get_seq <input>
this is easy to install on other systems as it seems to install all dependencies (click) by
simply writing pip install --editable.

The get_seq module is basically done. YAY!! The README requires more work and the --help
page is likely in need of greater clarity, bugs will likely arise overtime but it is good.